+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/flowsheet_limited_softening.py
@@ -13,30 +13,16 @@
 
 """Flowsheet examples that are limited (i.e. do not satisfy minimum viable product requirements)"""
 
-from pyomo.environ import (ConcreteModel, Objective, Expression, Constraint, Param,
-        TransformationFactory, value, units as pyunits)
-from pyomo.network import Arc
-from pyomo.util import infeasible
-from idaes.core import FlowsheetBlock
+from pyomo.environ import (Objective, Param)
 from idaes.core.util.scaling import (calculate_scaling_factors,
-                                     unscaled_constraints_generator,
-                                     unscaled_variables_generator,
-                                     badly_scaled_var_generator,
-                                     constraint_autoscale_large_jac)
-from idaes.core.util.initialization import propagate_state
+                                     badly_scaled_var_generator)
 from proteuslib.flowsheets.full_treatment_train.flowsheet_components import (pretreatment_softening,
-                                                                             desalination,
-                                                                             translator_block,
-                                                                             costing,
-                                                                             financials)
-from proteuslib.flowsheets.full_treatment_train.model_components import property_models
-from proteuslib.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
+                                                                             desalination)
 
 # Added import statements for testing.
 #       Need the pretreatment_stoich_softening_block functions to setup
 #       flowsheet to solve for lime dosage
-from idaes.core.util import scaling as iscale
-from proteuslib.flowsheets.full_treatment_train.chemistry_flowsheets.pretreatment_stoich_softening_block import *
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components.chemistry.pretreatment_stoich_softening_block import *
 
 def build_flowsheet_limited_softening(m, has_desal_feed=False, is_twostage=False, has_ERD=False,
                                       RO_type='Sep', RO_base='TDS', RO_level='simple'):
