+++ proteuslib/flowsheets/full_treatment_train/flowsheet_components/flowsheet_mvp_softening.py
@@ -11,27 +11,12 @@
 #
 ###############################################################################
 
-from pyomo.environ import ConcreteModel, Objective, Expression, Constraint, TransformationFactory, value, Param, Block
-from pyomo.environ import units as pyunits
-from pyomo.network import Arc
-from pyomo.util import infeasible
-from idaes.core import FlowsheetBlock
-from idaes.core.util.scaling import (calculate_scaling_factors,
-                                     constraint_autoscale_large_jac)
-from idaes.core.util.initialization import propagate_state
+from pyomo.environ import Objective, Param
+from idaes.core.util.scaling import (calculate_scaling_factors)
 from proteuslib.flowsheets.full_treatment_train.flowsheet_components import (desalination,
-                                                                             translator_block,
-                                                                             feed_block,
                                                                              gypsum_saturation_index,
-                                                                             costing,
-                                                                             financials,
                                                                              pretreatment_softening)
-from proteuslib.flowsheets.full_treatment_train.model_components import property_models
-from proteuslib.flowsheets.full_treatment_train.util import (solve_with_user_scaling,
-                                                             check_dof,
-                                                             check_build,
-                                                             check_scaling)
-from proteuslib.flowsheets.full_treatment_train.chemistry_flowsheets.pretreatment_stoich_softening_block import *
+from proteuslib.flowsheets.full_treatment_train.flowsheet_components.chemistry.pretreatment_stoich_softening_block import *
 
 """Flowsheet examples that satisfy minimum viable product requirements"""
 def build_flowsheet_mvp_softening(m, **kwargs):
